src/main/java/tudelft/caesarshift/CaesarShiftCipher.java
package tudelft.caesarshift;

import java.util.Objects;

public class CaesarShiftCipher {

    public static String caesarShiftCipher(String message, int shift){
        Objects.requireNonNull(message);
        StringBuilder sb = new StringBuilder();
        char currentChar;
        int length = message.length();

        shift = shift%26;

        for(int i = 0; i < length; i++){
            currentChar = message.charAt(i);
           
            //sb.append(currentChar);
            if (currentChar > 'z' || currentChar < 'a') {
                return "invalid";
            } else if ((char) (currentChar + shift) > 'z') {
                currentChar = (char) (currentChar - 26);
            } else if ((char) (currentChar + shift) < 'a'){
                currentChar = (char) (currentChar + 26);
            }
            sb.append((char) (currentChar + shift));
        }

        return sb.toString();
    }
}
